Revert "moved into package (test)"
This reverts commit a547eb3ba017be4c2761e954be8ca516f79e5298.
