+++ src/main/java/at/hannibal2/skyhanni/data/ElectionAPI.kt
@@ -1,9 +1,10 @@
-package at.hannibal2.skyhanni.data.mayor
+package at.hannibal2.skyhanni.data
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.config.ConfigManager
-import at.hannibal2.skyhanni.data.Candidate
-import at.hannibal2.skyhanni.data.Perk
+import at.hannibal2.skyhanni.data.Candidate.Companion.getMayorFromPerk
+import at.hannibal2.skyhanni.data.Candidate.Companion.setAssumeMayorJson
+import at.hannibal2.skyhanni.data.Perk.Companion.getPerkFromName
 import at.hannibal2.skyhanni.data.jsonobjects.other.MayorCandidate
 import at.hannibal2.skyhanni.data.jsonobjects.other.MayorElection
 import at.hannibal2.skyhanni.data.jsonobjects.other.MayorJson
@@ -27,6 +28,7 @@ import at.hannibal2.skyhanni.utils.RegexUtils.matches
 import at.hannibal2.skyhanni.utils.SimpleTimeMark
 import at.hannibal2.skyhanni.utils.SimpleTimeMark.Companion.asTimeMark
 import at.hannibal2.skyhanni.utils.SkyBlockTime
+import at.hannibal2.skyhanni.utils.SkyBlockTime.Companion.SKYBLOCK_YEAR_MILLIS
 import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import at.hannibal2.skyhanni.utils.json.fromJson
 import at.hannibal2.skyhanni.utils.repopatterns.RepoPattern
@@ -42,7 +44,7 @@ import kotlin.time.Duration.Companion.minutes
 @SkyHanniModule
 object ElectionAPI {
 
-    private val group = RepoPattern.Companion.group("mayorapi")
+    private val group = RepoPattern.group("mayorapi")
 
     /**
      * REGEX-TEST: Schedules an extra §bFishing Festival §7event during the year.
@@ -90,17 +92,17 @@ object ElectionAPI {
     var currentMinister: Candidate? = null
         private set
     private var lastMayor: Candidate? = null
-    var jerryExtraMayor: Pair<Candidate?, SimpleTimeMark> = null to SimpleTimeMark.Companion.farPast()
+    var jerryExtraMayor: Pair<Candidate?, SimpleTimeMark> = null to SimpleTimeMark.farPast()
         private set
-    private var lastJerryExtraMayorReminder = SimpleTimeMark.Companion.farPast()
+    private var lastJerryExtraMayorReminder = SimpleTimeMark.farPast()
 
-    private var lastUpdate = SimpleTimeMark.Companion.farPast()
+    private var lastUpdate = SimpleTimeMark.farPast()
     private val dispatcher = Dispatchers.IO
 
     private var rawMayorData: MayorJson? = null
     private var candidates = mapOf<Int, MayorCandidate>()
 
-    var nextMayorTimestamp = SimpleTimeMark.Companion.farPast()
+    var nextMayorTimestamp = SimpleTimeMark.farPast()
         private set
 
     private const val ELECTION_END_MONTH = 3 // Late Spring
@@ -110,7 +112,7 @@ object ElectionAPI {
      * @param input: The name of the mayor
      * @return: The NotEnoughUpdates color of the mayor; If no mayor was found, it will return "§c"
      */
-    fun mayorNameToColorCode(input: String): String = Candidate.Companion.getMayorFromName(input)?.color ?: "§c"
+    fun mayorNameToColorCode(input: String): String = Candidate.getMayorFromName(input)?.color ?: "§c"
 
     /**
      * @param input: The name of the mayor
@@ -129,17 +131,17 @@ object ElectionAPI {
         if (!LorenzUtils.inSkyBlock) return
         if (!Candidate.JERRY.isActive()) return
         if (jerryExtraMayor.first != null && jerryExtraMayor.second.isInPast()) {
-            jerryExtraMayor = null to SimpleTimeMark.Companion.farPast()
+            jerryExtraMayor = null to SimpleTimeMark.farPast()
             ChatUtils.clickableChat(
                 "The Perkpocalypse Mayor has expired! Click here to update the new temporary Mayor.",
                 onClick = { HypixelCommands.calendar() },
             )
         }
-        val misc = SkyHanniMod.Companion.feature.misc
+        val misc = SkyHanniMod.feature.misc
         if (jerryExtraMayor.first == null && misc.unknownPerkpocalypseMayorWarning) {
             if (lastJerryExtraMayorReminder.passedSince() < 5.minutes) return
             if (ReminderUtils.isBusy()) return
-            lastJerryExtraMayorReminder = SimpleTimeMark.Companion.now()
+            lastJerryExtraMayorReminder = SimpleTimeMark.now()
             ChatUtils.clickableChat(
                 "The Perkpocalypse Mayor is not known! Click here to update the temporary Mayor.",
                 onClick = { HypixelCommands.calendar() },
@@ -173,10 +175,9 @@ object ElectionAPI {
 
         val perk = stack.getLore().nextAfter({ perkpocalypsePerksPattern.matches(it) }, 2) ?: return
         // This is the first Perk of the Perkpocalypse Mayor
-        val jerryMayor =
-            Candidate.Companion.getMayorFromPerk(Perk.Companion.getPerkFromName(perk.removeColor()) ?: return)?.addAllPerks() ?: return
+        val jerryMayor = getMayorFromPerk(getPerkFromName(perk.removeColor()) ?: return)?.addAllPerks() ?: return
 
-        val lastMayorTimestamp = nextMayorTimestamp - SkyBlockTime.Companion.SKYBLOCK_YEAR_MILLIS.milliseconds
+        val lastMayorTimestamp = nextMayorTimestamp - SKYBLOCK_YEAR_MILLIS.milliseconds
 
         val expireTime = (1..21)
             .map { lastMayorTimestamp + (6.hours * it) }
@@ -200,7 +201,7 @@ object ElectionAPI {
     }
 
     private fun calculateNextMayorTime(): SimpleTimeMark {
-        val now = SkyBlockTime.Companion.now()
+        val now = SkyBlockTime.now()
 
         return SkyBlockTime(now.getElectionYear() + 1, ELECTION_END_MONTH, day = ELECTION_END_DAY).asTimeMark()
     }
@@ -214,12 +215,12 @@ object ElectionAPI {
             if (lastUpdate.passedSince() < 20.minutes) return
             if (currentMayor == Candidate.UNKNOWN && lastUpdate.passedSince() < 1.minutes) return
         }
-        lastUpdate = SimpleTimeMark.Companion.now()
+        lastUpdate = SimpleTimeMark.now()
 
-        SkyHanniMod.Companion.coroutineScope.launch {
+        SkyHanniMod.coroutineScope.launch {
             val url = "https://api.hypixel.net/v2/resources/skyblock/election"
             val jsonObject = withContext(dispatcher) { APIUtils.getJSONResponse(url) }
-            rawMayorData = ConfigManager.Companion.gson.fromJson<MayorJson>(jsonObject)
+            rawMayorData = ConfigManager.gson.fromJson<MayorJson>(jsonObject)
             val data = rawMayorData ?: return@launch
             val map = mutableMapOf<Int, MayorCandidate>()
             map put data.mayor.election.getPairs()
@@ -230,8 +231,8 @@ object ElectionAPI {
 
             val currentMayorName = data.mayor.name
             if (lastMayor?.name != currentMayorName) {
-                currentMayor = Candidate.Companion.setAssumeMayorJson(currentMayorName, data.mayor.perks)
-                currentMinister = data.mayor.minister?.let { Candidate.Companion.setAssumeMayorJson(it.name, listOf(it.perk)) }
+                currentMayor = setAssumeMayorJson(currentMayorName, data.mayor.perks)
+                currentMinister = data.mayor.minister?.let { setAssumeMayorJson(it.name, listOf(it.perk)) }
             }
         }
     }
@@ -242,7 +243,7 @@ object ElectionAPI {
 
     @SubscribeEvent
     fun onConfigReload(event: ConfigLoadEvent) {
-        val config = SkyHanniMod.Companion.feature.dev.debug.assumeMayor
+        val config = SkyHanniMod.feature.dev.debug.assumeMayor
         config.onToggle {
             val mayor = config.get()
